# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Matrimony API
  version: 0.1.0
  description: |
    The machine-readable API contract: every route, field and error code. The frontend
    generates its TypeScript types from this file (`npm run api:types` in the frontend repo),
    so keep it exact and keep it short.

    What each feature does, which endpoints it has and which frontend feature uses them is
    described phase by phase in the backend tracker, not here:
    https://claude.ai/artifact/RzGP4ky7jbKQMoUVA2WFYw

    Rules for every route: the agency comes from the request `Host` header (an allowlisted
    hostname; `X-Forwarded-Host` is ignored), except for `/health/*`. Every response is
    `cache-control: no-store`. Errors use `{ error: { code, requestId, details? } }`.
servers:
  - url: /
...
paths:
...
  /api/v1/public/tenant:
    get:
      tags: [Tenant]
      summary: Public configuration of the agency for this hostname
      description: Public. Branding and public content only, never secrets.
      responses:
        '200':
          description: The active agency for the request hostname.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/Error400' }
        '404': { $ref: '#/components/responses/TenantNotFound' }
...
  /api/v1/me:
    get:
      tags: [Account]
      summary: The signed-in account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The current tenant-local account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Self' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
...
  /api/v1/admin/staff/invitations:
    get:
      tags: [Staff]
      summary: Invitations that have not been accepted or cancelled
      description: |
        Admins only. Newest first. `status` is `expired` for an open invitation whose link no longer
        works; it can be sent again.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The open invitations of this agency.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffInvitationList' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
    post:
      tags: [Staff]
      summary: Invite a person to work as an agent or an admin
      description: |
        Admins only, 30 per 10 minutes by address, 30 an hour per admin and 3 an hour per invited
        address. Emails the link (valid 7 days, once). Inviting an address that already has an open
        invitation replaces it: a new link, the old one stops. An address that already has an
        account is 409 `EMAIL_IN_USE`; a member is never turned into staff by an invitation.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StaffInviteInput' }
      responses:
        '201':
          description: The invitation, now open.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffInvitation' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '409':
          description: EMAIL_IN_USE.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/admin/staff/invitations/{invitationId}/resend:
    post:
      tags: [Staff]
      summary: Send an open invitation again with a new link
      description: |
        Admins only. The earlier link stops working and the expiry starts again. 3 an hour per
        invitation. 404 `INVITATION_NOT_FOUND` when it was accepted, cancelled or is unknown.
      security:
        - bearerAuth: []
      parameters:
        - { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The invitation with its new expiry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffInvitation' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404':
          description: INVITATION_NOT_FOUND.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: EMAIL_IN_USE (the address got an account since).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/admin/staff/invitations/{invitationId}:
    delete:
      tags: [Staff]
      summary: Cancel an open invitation
      description: |
        Admins only. Its link stops working at once. The record is kept. 404
        `INVITATION_NOT_FOUND` when it was accepted, already cancelled or is unknown.
      security:
        - bearerAuth: []
      parameters:
        - { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: Cancelled.
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404':
          description: INVITATION_NOT_FOUND.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The short-lived access token from sign-in or refresh (ES256). Keep it in memory only.
...
  schemas:
    Tenant:
      type: object
      required: [id, name, locale, content]
      properties:
        id: { type: string }
        name: { type: string }
        locale: { type: string, enum: [bn, en] }
        content: { $ref: '#/components/schemas/PublicConfig' }
    LocalizedText:
      type: object
      description: Text in one or both languages. A reader sees their own language, else the other.
      properties:
        bn: { type: string, maxLength: 2000 }
        en: { type: string, maxLength: 2000 }
    PublicConfig:
      type: object
      description: The agency's non-secret public settings. Every field is optional; unknown fields are ignored.
      properties:
        name:
          type: string
          description: Display name of the agency. Falls back to the agency's `name`.
        branding:
          type: object
          properties:
            logoUrl:
              type: string
              description: An `https://` URL or a path starting with `/`. Anything else is ignored.
            primaryColor:
              type: string
              pattern: '^#[0-9a-fA-F]{6}$'
              description: Main brand color as `#RRGGBB`. It reaches CSS, so nothing else is accepted.
        contact:
          type: object
          properties:
            phone: { type: string, maxLength: 40 }
            email: { type: string, maxLength: 200 }
        home:
          type: object
          description: Optional homepage copy that replaces the platform's default wording.
          properties:
            title: { $ref: '#/components/schemas/LocalizedText' }
            subtitle: { $ref: '#/components/schemas/LocalizedText' }
        about: { $ref: '#/components/schemas/LocalizedText' }
        branches:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/Branch' }
        successStories:
          type: array
          maxItems: 6
          items: { $ref: '#/components/schemas/SuccessStory' }
    Branch:
      type: object
      required: [name]
      properties:
        name: { $ref: '#/components/schemas/LocalizedText' }
        address: { $ref: '#/components/schemas/LocalizedText' }
        phone: { type: string, maxLength: 40 }
        mapUrl:
          type: string
          description: An `https://` link to the location on a map. Anything else is ignored.
    SuccessStory:
      type: object
      required: [names]
      description: Only stories the couple has agreed to publish belong here.
      properties:
        names: { $ref: '#/components/schemas/LocalizedText' }
        story: { $ref: '#/components/schemas/LocalizedText' }
        year: { type: integer, minimum: 1900, maximum: 2100 }
    Self:
      type: object
      required: [id, agencyId, role, displayName]
      properties:
        id: { type: string }
        agencyId: { type: string }
        role: { type: string, enum: [admin, agent, member] }
        displayName: { type: string }
...
    StaffInviteInput:
      type: object
      additionalProperties: false
      required: [email, displayName, role, locale]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        role: { type: string, enum: [agent, admin] }
        locale:
          type: string
          enum: [bn, en]
          description: The language of the email and of the page its link opens.
    StaffInvitation:
      type: object
      required: [id, email, displayName, role, locale, invitedByName, createdAt, expiresAt, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [agent, admin] }
        locale: { type: string, enum: [bn, en] }
        invitedByName: { type: string, description: The display name of the admin who invited. }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        status:
          type: string
          enum: [pending, expired]
          description: '`expired` is an open invitation whose link no longer works.'
    StaffInvitationList:
      type: object
      required: [invitations]
      properties:
        invitations:
          type: array
          items: { $ref: '#/components/schemas/StaffInvitation' }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, requestId]
          properties:
            code:
              type: string
              description: |
                Stable machine-readable code. Known codes: HOST_INVALID, TENANT_NOT_FOUND,
                BEARER_TOKEN_REQUIRED, INVALID_ACCESS_TOKEN, SESSION_INVALID, SESSION_EXPIRED,
                ACCOUNT_NOT_ACTIVE, ROLE_FORBIDDEN, ORIGIN_INVALID, CSRF_INVALID,
                REFRESH_UNAVAILABLE,
                INVALID_REQUEST, RATE_LIMITED, REQUEST_REJECTED, INTERNAL_ERROR, and for My profile:
                PROFILE_NOT_FOUND, PROFILE_VERSION_CONFLICT, PROFILE_LOCKED,
                PROFILE_EDIT_REQUIRES_REVIEW, PROFILE_STATE_INVALID, PROFILE_INCOMPLETE,
                NO_PENDING_REVIEW, NO_CHANGES, MEMBER_CODE_UNAVAILABLE, and for photos:
                PHOTO_NOT_FOUND, PHOTO_LIMIT_REACHED, PHOTO_NOT_APPROVED, PHOTO_INVALID,
                PHOTO_TYPE_UNSUPPORTED, PHOTO_TOO_SMALL, PAYLOAD_TOO_LARGE,
                UNSUPPORTED_MEDIA_TYPE, and for signing in and email links: INVALID_CREDENTIALS,
                TOO_MANY_ATTEMPTS (with details.retryAfter, seconds), EMAIL_RATE_LIMITED (with
                details.retryAfter, seconds), LINK_INVALID_OR_EXPIRED, and
                for Google: GOOGLE_NOT_CONFIGURED, GOOGLE_UNAVAILABLE, GOOGLE_AUTH_FAILED,
                GOOGLE_EMAIL_UNVERIFIED (Google has not confirmed the email), ACCOUNT_HAS_NO_PASSWORD (the account has no
                password to approve a link with), GOOGLE_ALREADY_LINKED, GOOGLE_RETRY, OAUTH_CHALLENGE_EXPIRED,
                OAUTH_CONTEXT_MISMATCH, and
                for adding an email: EMAIL_ALREADY_SET, PHONE_REQUIRED, EMAIL_IN_USE, and
                for inviting staff: INVITATION_NOT_FOUND (and EMAIL_IN_USE, LINK_INVALID_OR_EXPIRED), and
                for phone codes: PHONE_NOT_CONFIGURED, PHONE_COUNTRY_NOT_SUPPORTED, PHONE_NOT_REGISTERED, CODE_RATE_LIMITED (with details.retryAfter,
                seconds), SMS_UNAVAILABLE, CODE_INVALID, CODE_EXPIRED, PHONE_RETRY, PHONE_IN_USE, and
                for staff: REVIEW_NOT_FOUND, REVIEW_ALREADY_DECIDED, REVIEW_CANCELLED,
                REVIEW_OUT_OF_DATE, REVIEW_OWN_SUBMISSION, PROFILE_SELF_SERVICE,
                STATUS_CHANGE_NOT_ALLOWED, AGENT_NOT_AVAILABLE.
            details:
              type: object
              description: |
                Present on INVALID_REQUEST and PROFILE_INCOMPLETE. Each problem names a field by its
                dotted path (for example `profile.dateOfBirth`) and a stable code the client
                translates: required, tooLong, tooShort, tooWeak, sameAsEmail, invalid, invalidOption,
                invalidDate, underAge, tooOld, outOfRange, invalidPhone, invalidEmail, rangeOrder. Values are never echoed back.
              properties:
                fields:
                  type: array
                  items:
                    type: object
                    required: [path, code]
                    properties:
                      path: { type: string }
                      code: { type: string }
            requestId: { type: string, description: Quote this when reporting a problem. }
  responses:
    Error400:
      description: Invalid host or request (HOST_INVALID, INVALID_REQUEST).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error401:
      description: Not authenticated or session invalid. Sends `WWW-Authenticate Bearer`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error403:
      description: Forbidden (ROLE_FORBIDDEN, ACCOUNT_NOT_ACTIVE, ORIGIN_INVALID, CSRF_INVALID).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TenantNotFound:
      description: The hostname is not a provisioned, active agency (TENANT_NOT_FOUND).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests (RATE_LIMITED).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }